refactor: move smart-search into features dir
