+++ packages/obsidian-plugin/src/features/core/index.ts
@@ -1,21 +1,8 @@
-import { type } from "arktype";
-import type { Request, Response } from "express";
 import { Notice, Plugin, PluginSettingTab } from "obsidian";
-import { shake } from "radash";
 import { BehaviorSubject, firstValueFrom, lastValueFrom } from "rxjs";
-import {
-  jsonSearchRequest,
-  searchParameters,
-  type SearchResponse,
-} from "shared";
 import { mount, unmount } from "svelte";
 
-import {
-  loadLocalRestAPI,
-  loadSmartSearchAPI,
-  logger,
-  type Dependencies,
-} from "$/shared";
+import { loadLocalRestAPI, logger, type Dependencies } from "$/shared";
 
 import SettingsContainer from "./components/SettingsContainer.svelte";
 import type { McpToolsPluginSettings } from "./types";
@@ -23,6 +10,7 @@ import type { McpToolsPluginSettings } from "./types";
 import { setup as setupMcpServerInstall } from "../mcp-server-install";
 import { setup as setupSourceDocuments } from "../source-documents";
 import { setup as setupTemplates } from "../templates";
+import { setup as setupSmartSearch } from "../smart-search";
 
 export class McpToolsSettingTab extends PluginSettingTab {
   plugin: McpToolsPlugin;
@@ -59,13 +47,13 @@ export class McpToolsPlugin extends Plugin {
     installed: false,
   });
 
-  async onload() {
+  onload() {
     logger.info("Loading MCP Tools Plugin");
 
     // Add settings tab to plugin
     this.addSettingTab(new McpToolsSettingTab(this));
 
-    // Check for required dependencies
+    // Wait for required dependencies
     lastValueFrom(loadLocalRestAPI(this)).then(async (localRestApi) => {
       if (!localRestApi.api) {
         new Notice(
@@ -81,11 +69,7 @@ export class McpToolsPlugin extends Plugin {
       await setupMcpServerInstall(this);
       await setupSourceDocuments(this);
       await setupTemplates(this);
-
-      // Register endpoints
-      localRestApi.api
-        .addRoute("/search/smart")
-        .post(this.handleSearchRequest.bind(this));
+      await setupSmartSearch(this);
 
       logger.info("MCP Tools Plugin loaded");
     });
@@ -103,69 +87,6 @@ export class McpToolsPlugin extends Plugin {
     const localRestApi = await firstValueFrom(this.localRestApi$);
     // The API key is stored in the plugin's settings
     return localRestApi.plugin?.settings?.apiKey;
-  }
-
-  private async handleSearchRequest(req: Request, res: Response) {
-    try {
-      const dep = await lastValueFrom(loadSmartSearchAPI(this));
-      const smartSearch = dep.api;
-      if (!smartSearch) {
-        new Notice(
-          "Smart Search REST API Plugin: smart-connections plugin is required but not found. Please install it from the community plugins.",
-          0,
-        );
-        res.status(503).json({
-          error: "Smart Connections plugin is not available",
-        });
-        return;
-      }
-
-      // Validate request body
-      const requestBody = jsonSearchRequest
-        .pipe(({ query, filter = {} }) => ({
-          query,
-          filter: shake({
-            key_starts_with_any: filter.folders,
-            exclude_key_starts_with_any: filter.excludeFolders,
-            limit: filter.limit,
-          }),
-        }))
-        .to(searchParameters)(req.body);
-      if (requestBody instanceof type.errors) {
-        res.status(400).json({
-          error: "Invalid request body",
-          summary: requestBody.summary,
-        });
-        return;
-      }
-
-      // Perform search
-      const results = await smartSearch.search(
-        requestBody.query,
-        requestBody.filter,
-      );
-
-      // Format response
-      const response: SearchResponse = {
-        results: await Promise.all(
-          results.map(async (result) => ({
-            path: result.item.path,
-            text: await result.item.read(),
-            score: result.score,
-            breadcrumbs: result.item.breadcrumbs,
-          })),
-        ),
-      };
-
-      res.json(response);
-      return;
-    } catch (error) {
-      logger.error("Smart Search API error:", { error, body: req.body });
-      res.status(503).json({
-        error: "An error occurred while processing the search request",
-      });
-      return;
-    }
   }
 
   onunload() {
